+++ esphome/components/display/bufferex_1bit_2color.h
@@ -12,9 +12,9 @@ class Bufferex1bit2color : public display::BufferexBase {
   void set_buffer(int x, int y, Color color) override;
   void fill_buffer(Color color) override;
   uint16_t get_pixel_to_565(int x, int y) override;
-  uint16_t get_pixel_to_565(uint16_t pos) override;
+  uint16_t get_pixel_to_565(uint32_t pos) override;
   uint32_t get_pixel_to_666(int x, int y) override;
-  uint32_t get_pixel_to_666(uint16_t pos) override;
+  uint32_t get_pixel_to_666(uint32_t pos) override;
   size_t get_buffer_length() override;
   size_t get_buffer_size() override;
 
@@ -24,10 +24,10 @@ class Bufferex1bit2color : public display::BufferexBase {
  protected:
   Color color_on_ = COLOR_WHITE;
   Color color_off_ = COLOR_BLACK;
-  uint16_t get_pixel_buffer_position_internal_(int x, int y);
+  uint32_t get_pixel_buffer_position_internal_(int x, int y);
 
   uint8_t get_color_bit_(int x, int y);
-  uint8_t get_color_bit_(uint16_t pos);
+  uint8_t get_color_bit_(uint32_t pos);
   display::BufferType get_buffer_type() override { return this->buffer_type_; }
   uint8_t get_pixel_storage_size() override { return this->pixel_storage_size_; }
