+++ esphome/components/display/bufferex_332.cpp
@@ -25,7 +25,7 @@ void Bufferex332::fill_buffer(Color color) {
 }
 
 void HOT Bufferex332::set_buffer(int x, int y, Color color) {
-  uint16_t pos = get_pixel_buffer_position_(x, y);
+  uint32_t pos = get_pixel_buffer_position_(x, y);
   const uint16_t color332 = color.to_332();
   this->buffer_[pos] = color332;
 }
@@ -33,13 +33,13 @@ void HOT Bufferex332::set_buffer(int x, int y, Color color) {
 size_t Bufferex332::get_buffer_size() { return this->get_buffer_length(); }
 
 // 565
-uint16_t Bufferex332::get_pixel_to_565(uint16_t pos) {
+uint16_t Bufferex332::get_pixel_to_565(uint32_t pos) {
   return Color(this->buffer_[pos], Color::ColorOrder::COLOR_ORDER_RGB, Color::ColorBitness::COLOR_BITNESS_332, true)
       .to_565();
 }
 
 // 666
-uint32_t Bufferex332::get_pixel_to_666(uint16_t pos) {
+uint32_t Bufferex332::get_pixel_to_666(uint32_t pos) {
   return Color(this->buffer_[pos], Color::ColorOrder::COLOR_ORDER_RGB, Color::ColorBitness::COLOR_BITNESS_565, true)
       .to_666();
 }
