+++ esphome/components/display/bufferex_indexed8.h
@@ -13,9 +13,9 @@ class BufferexIndexed8 : public display::BufferexBase {
 
   void fill_buffer(Color color) override;
   uint16_t get_pixel_to_565(int x, int y) override;
-  uint16_t get_pixel_to_565(uint16_t pos) override;
+  uint16_t get_pixel_to_565(uint32_t pos) override;
   uint32_t get_pixel_to_666(int x, int y) override;
-  uint32_t get_pixel_to_666(uint16_t pos) override;
+  uint32_t get_pixel_to_666(uint32_t pos) override;
   size_t get_buffer_length() override;
   size_t get_buffer_size() override;
   void set_index_size(uint8_t index_size) { this->index_size_ = index_size; }
@@ -29,7 +29,7 @@ class BufferexIndexed8 : public display::BufferexBase {
   size_t get_buffer_length_ = 0;
 
   uint8_t get_index_value_(int x, int y);
-  uint8_t get_index_value_(uint16_t pos);
+  uint8_t get_index_value_(uint32_t pos);
   display::BufferType get_buffer_type() override { return this->buffer_type_; }
   uint8_t get_pixel_storage_size() override { return this->pixel_storage_size_; }
